+++ connector/core/core-service/src/main/java/org/oaebudt/edc/core/store/InMemoryParticipantGroupStore.java
@@ -1,5 +1,6 @@
 package org.oaebudt.edc.core.store;
 
+import org.oaebudt.edc.spi.store.ParticipantGroup;
 import org.oaebudt.edc.spi.store.ParticipantGroupStore;
 
 import java.util.Arrays;
@@ -10,24 +11,24 @@ import java.util.Set;
 
 public class InMemoryParticipantGroupStore implements ParticipantGroupStore {
 
-    private Map<String, Set<String>> participantGroups = new HashMap<>();
+    private final Map<String, ParticipantGroup> participantGroups = new HashMap<>();
 
     @Override
-    public Set<String> getParticipantsByGroupId(String groupId) {
+    public ParticipantGroup findById(String groupId) {
         return participantGroups.get(groupId);
     }
 
     @Override
-    public void putParticipantsInGroup(String groupId, String... participantIds) {
+    public void save(String groupId, String... participantIds) {
         if(participantGroups.containsKey(groupId)) {
-            participantGroups.get(groupId).addAll(Arrays.asList(participantIds));
+            participantGroups.get(groupId).participants().addAll(Arrays.asList(participantIds));
             return;
         }
-        participantGroups.put(groupId, new HashSet<>(Arrays.asList(participantIds)));
+        participantGroups.put(groupId, new ParticipantGroup(groupId, new HashSet<>(Arrays.asList(participantIds))));
     }
 
     @Override
-    public Map<String, Set<String>> getAllGroups() {
-        return participantGroups;
+    public Set<ParticipantGroup> getAllGroups() {
+        return new HashSet<>(participantGroups.values());
     }
 }
