--- connector/extensions/report-service/src/main/java/org/oaebudt/edc/report/model/Constants.java
@@ -1,6 +1,0 @@
-package org.oaebudt.edc.report.model;
-
-public class Constants {
-
-    public static final String DEFAULT_POLICY_ID = "require-membership";
-}
