--- connector/extensions/report-service/src/main/java/org/oaebudt/edc/report/model/ReportType.java
@@ -1,5 +1,0 @@
-package org.oaebudt.edc.report.model;
-
-public enum ReportType {
-    TITLE_REPORT, ITEM_REPORT, TEST_REPORT
-}
