+++ connector/extensions/web-service/src/main/java/org/oaebudt/edc/web/ConsumerApiController.java
@@ -1,0 +1,35 @@
+package org.oaebudt.edc.web;
+
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import org.bson.Document;
+import org.oaebudt.edc.web.model.ReportType;
+import org.oaebudt.edc.web.repository.ReportStore;
+
+import static jakarta.ws.rs.core.MediaType.WILDCARD;
+
+@Path("/")
+@Consumes(WILDCARD)
+@Produces(WILDCARD)
+public class ConsumerApiController {
+    private final ReportStore reportStore;
+
+    public ConsumerApiController(ReportStore reportStore) {
+        this.reportStore = reportStore;
+    }
+
+    @GET
+    @Path("report")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response getReportWithSearchCriteria(@QueryParam("reportType") ReportType reportType) {
+        Document document = reportStore.getReportByType(reportType);
+        return Response.status(Response.Status.CREATED)
+                .entity(document)
+                .build();
+    }
+}
