+++ connector/extensions/web-service/src/main/java/org/oaebudt/edc/web/dto/AddParticipantToGroupRequest.java
@@ -1,0 +1,6 @@
+package org.oaebudt.edc.web.dto;
+
+import java.util.Set;
+
+public record AddParticipantToGroupRequest(String groupName, Set<String> participants) {
+}
