+++ connector/extensions/web-service/src/main/java/org/oaebudt/edc/web/model/Constants.java
@@ -1,0 +1,6 @@
+package org.oaebudt.edc.web.model;
+
+public class Constants {
+
+    public static final String DEFAULT_POLICY_ID = "require-membership";
+}
