+++ connector/extensions/web-service/src/main/java/org/oaebudt/edc/web/model/ReportType.java
@@ -1,0 +1,5 @@
+package org.oaebudt.edc.web.model;
+
+public enum ReportType {
+    TITLE_REPORT, ITEM_REPORT
+}
