+++ connector/extensions/web-service/src/main/java/org/oaebudt/edc/web/repository/ReportStore.java
@@ -1,0 +1,10 @@
+package org.oaebudt.edc.web.repository;
+
+import org.bson.Document;
+import org.oaebudt.edc.web.model.ReportType;
+
+public interface ReportStore {
+    void saveReport(String json, ReportType reportType);
+
+    Document getReportByType(ReportType reportType);
+}
