+++ connector/extensions/web-service/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,0 +1,1 @@
+org.oaebudt.edc.web.WebApiExtension
