+++ connector/launchers/runtime-embedded/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 
 dependencies {
     runtimeOnly(project(":core:core-service"))
-    runtimeOnly(project(":extensions:report-service"))
+    runtimeOnly(project(":extensions:web-service"))
     runtimeOnly(project(":extensions:keycloak-auth"))
     runtimeOnly(project(":extensions:catalog-node-resolver"))
     runtimeOnly(project(":extensions:dcp-impl"))
