+++ connector/settings.gradle.kts
@@ -26,7 +26,7 @@ include(":extensions:superuser-seed")
 include(":extensions:dcp-impl")
 include(":extensions:catalog-node-resolver")
 include(":extensions:credential-from-file")
-include(":extensions:report-service")
+include(":extensions:web-service")
 
 include(":launchers:controlplane")
 include(":launchers:dataplane")
