+++ connector/spi/core-spi/src/main/java/org/oaebudt/edc/spi/store/ParticipantGroup.java
@@ -1,0 +1,6 @@
+package org.oaebudt.edc.spi.store;
+
+import java.util.Set;
+
+public record ParticipantGroup(String id, Set<String> participants) {
+}
