+++ connector/spi/core-spi/src/main/java/org/oaebudt/edc/spi/store/ParticipantGroupStore.java
@@ -1,13 +1,12 @@
 package org.oaebudt.edc.spi.store;
 
-import java.util.Map;
 import java.util.Set;
 
 public interface ParticipantGroupStore {
 
-    Set<String> getParticipantsByGroupId(String groupId);
+    ParticipantGroup findById(String groupId);
 
-    void putParticipantsInGroup(String groupId, String... participantIds);
+    void save(String groupId, String... participantIds);
 
-    Map<String, Set<String>> getAllGroups();
+    Set<ParticipantGroup> getAllGroups();
 }
