+++ connector/tests/src/test/java/org/oaebudt/edc/utils/OaebudtParticipant.java
@@ -63,14 +63,11 @@ public class OaebudtParticipant extends Participant {
     protected LazySupplier<URI> identityHubSts = new LazySupplier<>(() ->
             URI.create("http://localhost:" + Ports.getFreePort() + "/api/sts"));
 
-    protected LazySupplier<URI> reportServiceUrl = new LazySupplier<>(() ->
+    protected LazySupplier<URI> webServiceUrl = new LazySupplier<>(() ->
             URI.create("http://localhost:" + Ports.getFreePort() + "/api/report/"));
 
     protected LazySupplier<URI> consumerServiceUrl = new LazySupplier<>(() ->
             URI.create("http://localhost:" + Ports.getFreePort() + "/api/consumer/"));
-
-    protected LazySupplier<URI> participantServiceUrl = new LazySupplier<>(() ->
-            URI.create("http://localhost:" + Ports.getFreePort() + "/api/participant/"));
 
     private OaebudtParticipant() {}
 
@@ -126,14 +123,11 @@ public class OaebudtParticipant extends Participant {
                 entry("web.http.public.port", getFreePort() + ""),
                 entry("web.http.version.path", "/version"),
                 entry("web.http.version.port", getFreePort() + ""),
-                entry("web.http.report.port", reportServiceUrl.get().getPort() + ""),
-                entry("web.http.report.path", reportServiceUrl.get().getPath()),
-                entry("web.http.report.auth.type", "keycloak"),
+                entry("web.http.web.port", webServiceUrl.get().getPort() + ""),
+                entry("web.http.web.path", webServiceUrl.get().getPath()),
+                entry("web.http.web.auth.type", "keycloak"),
                 entry("web.http.consumer.port", consumerServiceUrl.get().getPort() + ""),
-                entry("web.http.consumer.path", consumerServiceUrl.get().getPath()),
-                entry("web.http.participant.port", participantServiceUrl.get().getPort() + ""),
-                entry("web.http.participant.path", participantServiceUrl.get().getPath()),
-                entry("web.http.participant.auth.type", "keycloak")
+                entry("web.http.consumer.path", consumerServiceUrl.get().getPath())
         );
 
         return ConfigFactory.fromMap(map);
@@ -186,11 +180,8 @@ public class OaebudtParticipant extends Participant {
         return identityHubSts.get().getPort();
     }
 
-    public LazySupplier<URI> getReportServiceUrl() {
-        return reportServiceUrl;
-    }
-    public LazySupplier<URI> getParticipantServiceUrl() {
-        return participantServiceUrl;
+    public LazySupplier<URI> getWebServiceUrl() {
+        return webServiceUrl;
     }
 
     public RequestSpecification baseManagementRequest() {
